#21 작업중 코드 정리
